# GreekLearningApp-Frontend/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Greek Parser API",
    "description": "API for Greek language learning and parsing",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5001"
    }
  ],
  "paths": {
...
    "/api/translations": {
      "post": {
        "tags": [
          "Translations"
        ],
        "operationId": "Translations_CreateTranslation",
        "requestBody": {
          "x-name": "translation",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimpleTranslationDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleTranslationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{bookId}/translations": {
      "get": {
        "tags": [
          "Translations"
        ],
        "operationId": "Translations_GetByBookId",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TranslationDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Translations"
        ],
        "operationId": "Translations_Create",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "createDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTranslationDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/translations/{id}": {
      "get": {
        "tags": [
          "Translations"
        ],
        "operationId": "Translations_GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Translations"
        ],
        "operationId": "Translations_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "updateDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTranslationDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Translations"
        ],
        "operationId": "Translations_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/translations/{translationId}/units": {
      "get": {
        "tags": [
          "TranslationUnits"
        ],
        "operationId": "TranslationUnits_GetByTranslationId",
        "parameters": [
          {
            "name": "translationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TranslationUnitDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TranslationUnits"
        ],
        "operationId": "TranslationUnits_Create",
        "parameters": [
          {
            "name": "translationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "createDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTranslationUnitDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationUnitDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/translation-units/{id}": {
      "get": {
        "tags": [
          "TranslationUnits"
        ],
        "operationId": "TranslationUnits_GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationUnitDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TranslationUnits"
        ],
        "operationId": "TranslationUnits_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "updateDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTranslationUnitDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationUnitDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TranslationUnits"
        ],
        "operationId": "TranslationUnits_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/translations/{translationId}/units/batch": {
      "post": {
        "tags": [
          "TranslationUnits"
        ],
        "operationId": "TranslationUnits_BatchCreate",
        "parameters": [
          {
            "name": "translationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "batchDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreateTranslationUnitsDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TranslationUnitDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SimpleTranslationDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "translationGuid": {
            "type": "string",
            "format": "guid"
          },
          "translationId": {
            "type": "integer",
            "format": "int32"
          },
          "unitId": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "string"
          }
        }
      },
      "TranslationDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "bookId": {
            "type": "integer",
            "format": "int32"
          },
          "languageCode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateTranslationDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "languageCode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "UpdateTranslationDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "languageCode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "TranslationUnitDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "translationId": {
            "type": "integer",
            "format": "int32"
          },
          "chapterId": {
            "type": "integer",
            "format": "int32"
          },
          "originalUnitPath": {
            "type": "string"
          },
          "translatedText": {
            "type": "string"
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "scopeStart": {
            "type": "string",
            "nullable": true
          },
          "scopeEnd": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreateTranslationUnitDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "chapterId": {
            "type": "integer",
            "format": "int32"
          },
          "originalUnitPath": {
            "type": "string"
          },
          "translatedText": {
            "type": "string"
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "scopeStart": {
            "type": "string",
            "nullable": true
          },
          "scopeEnd": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "BatchCreateTranslationUnitsDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "units": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTranslationUnitDto"
            }
          }
        }
      },
      "UpdateTranslationUnitDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "chapterId": {
            "type": "integer",
            "format": "int32"
          },
          "originalUnitPath": {
            "type": "string"
          },
          "translatedText": {
            "type": "string"
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "scopeStart": {
            "type": "string",
            "nullable": true
          },
          "scopeEnd": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}